05-google-cloud-platform/01-bigquery/requirements.txt
google-cloud==0.34.0
pandas-gbq==0.24.0
python-dotenv==1.0.1